+++ sulong/projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/CommonNodeFactory.java
@@ -131,6 +131,7 @@ import com.oracle.truffle.llvm.runtime.nodes.literals.LLVMSimpleLiteralNodeFacto
 import com.oracle.truffle.llvm.runtime.nodes.literals.LLVMSimpleLiteralNodeFactory.LLVMI8LiteralNodeGen;
 import com.oracle.truffle.llvm.runtime.nodes.literals.LLVMSimpleLiteralNodeFactory.LLVMIVarBitLiteralNodeGen;
 import com.oracle.truffle.llvm.runtime.nodes.literals.LLVMSimpleLiteralNodeFactory.LLVMNativePointerLiteralNodeGen;
+import com.oracle.truffle.llvm.runtime.nodes.memory.LLVMVectorizedGetElementPtrNodeGen;
 import com.oracle.truffle.llvm.runtime.nodes.memory.load.LLVMDirectLoadNode.LLVMPointerDirectLoadNode;
 import com.oracle.truffle.llvm.runtime.nodes.memory.load.LLVMDirectLoadNodeFactory.LLVM80BitFloatDirectLoadNodeGen;
 import com.oracle.truffle.llvm.runtime.nodes.memory.load.LLVMDirectLoadNodeFactory.LLVMIVarBitDirectLoadNodeGen;
@@ -225,6 +226,7 @@ import com.oracle.truffle.llvm.runtime.nodes.vars.LLVMReadVectorNodeFactory.LLVM
 import com.oracle.truffle.llvm.runtime.nodes.vars.LLVMReadVectorNodeFactory.LLVMI64VectorReadNodeGen;
 import com.oracle.truffle.llvm.runtime.nodes.vars.LLVMReadVectorNodeFactory.LLVMI8VectorReadNodeGen;
 import com.oracle.truffle.llvm.runtime.pointer.LLVMNativePointer;
+import com.oracle.truffle.llvm.runtime.types.AggregateType;
 import com.oracle.truffle.llvm.runtime.types.ArrayType;
 import com.oracle.truffle.llvm.runtime.types.FunctionType;
 import com.oracle.truffle.llvm.runtime.types.MetaType;
@@ -437,6 +439,124 @@ public class CommonNodeFactory {
             return null;
         } else {
             return LLVMDebugInitNodeFactory.AggregateInitNodeGen.create(targetSlot, offsets, lengths);
+        }
+    }
+
+    /**
+     * The rules for whether to build a scalar-getelementptr or vector-getelementptr node:
+     *
+     * <pre>
+     * S = scalar node
+     * V = vector node
+     * BC = broadcast node
+     * GEP = scalar getelementptr node
+     * VGEP = vector getelementptr node
+     *
+     * The BasePointer (BP) could either be a scalar, vector or GEP/VGEP node coming from the
+     * previous indexing dimension. The Index can only either be a scalar or a vector.
+     *
+     * Pointers, arrays and structures are considered scalars.
+     *
+     * (BP, Idx) --> Next OP(PTR, IDX)
+     *
+     * -------------------------------
+     *
+     * 0: (S, S) --> GEP(S, S)
+     * 1: (S, V) --> VGEP(BC(S), V)
+     * 2: (V, S) --> VGEP(V, BC(S))
+     * 3: (V, V) --> VGEP(V, V)
+     * 4: (GEP, S) --> GEP(GEP, S)
+     * 5: (GEP, V) --> VGEP(BC(GEP), V)
+     * 6: (VGEP, S) --> VGEP(VGEP, BC(S))
+     * 7: (VGEP, V) --> VGEP(VGEP, V)
+     * </pre>
+     */
+    public static LLVMExpressionNode createElementPointer(NodeFactory nodeFactory, long indexedTypeLength, Type currentType, LLVMExpressionNode currentAddress, LLVMExpressionNode indexNode,
+                    Type indexType,
+                    final boolean wasVectorized) {
+        if (wasVectorized) {
+            // Cases 2, 3, 6, 7
+            if (indexType instanceof VectorType) {
+                // Cases 3, 7
+                return nodeFactory.createVectorizedTypedElementPointer(indexedTypeLength, currentType, currentAddress, indexNode);
+            } else {
+                // Cases 2, 6
+                int length = ((VectorType) currentType).getNumberOfElementsInt();
+                return nodeFactory.createVectorizedTypedElementPointer(indexedTypeLength, currentType, currentAddress,
+                                LLVMVectorizedGetElementPtrNodeGen.IndexVectorBroadcastNodeGen.create(length, indexNode));
+            }
+        } else {
+            // Cases 0, 1, 4, 5
+            if (indexType instanceof VectorType) {
+                // Cases 1, 5
+                int length = ((VectorType) indexType).getNumberOfElementsInt();
+                return nodeFactory.createVectorizedTypedElementPointer(indexedTypeLength, currentType, LLVMVectorizedGetElementPtrNodeGen.ResultVectorBroadcastNodeGen.create(length, currentAddress),
+                                indexNode);
+            } else {
+                // Cases 0, 4
+                return nodeFactory.createTypedElementPointer(indexedTypeLength, currentType, currentAddress, indexNode);
+            }
+        }
+    }
+
+    public static LLVMExpressionNode createNestedElementPointerNode(
+                    NodeFactory nodeFactory,
+                    DataLayout dataLayout,
+                    LLVMExpressionNode[] indexNodes,
+                    Long[] indexConstants,
+                    Type[] indexTypes,
+                    LLVMExpressionNode curAddress,
+                    Type curType) {
+        try {
+            LLVMExpressionNode currentAddress = curAddress;
+            Type currentType = curType;
+
+            boolean wasVectorized = currentType instanceof VectorType;
+            if (wasVectorized) {
+                VectorType vectorType = (VectorType) currentType;
+                currentType = vectorType.getElementType();
+            }
+
+            for (int i = 0; i < indexNodes.length; i++) {
+                Type indexType = indexTypes[i];
+                Long indexInteger = indexConstants[i];
+                if (indexInteger == null) {
+                    // the index is determined at runtime
+                    if (currentType instanceof StructureType) {
+                        // according to http://llvm.org/docs/LangRef.html#getelementptr-instruction
+                        throw new LLVMParserException("Indices on structs must be constant integers!");
+                    }
+                    AggregateType aggregate = (AggregateType) currentType;
+                    long indexedTypeLength = aggregate.getOffsetOf(1, dataLayout);
+                    currentType = aggregate.getElementType(1);
+                    currentAddress = createElementPointer(nodeFactory, indexedTypeLength, currentType, currentAddress, indexNodes[i], indexType, wasVectorized);
+                    wasVectorized = currentAddress instanceof LLVMVectorizedGetElementPtrNodeGen;
+                } else {
+                    // the index is a constant integer
+                    AggregateType aggregate = (AggregateType) currentType;
+                    long addressOffset = aggregate.getOffsetOf(indexInteger, dataLayout);
+                    currentType = aggregate.getElementType(indexInteger);
+
+                    // creating a pointer inserts type information, this needs to happen for the
+                    // address computed by getelementptr even if it is the same as the basepointer
+                    if (addressOffset != 0 || i == indexNodes.length - 1) {
+                        LLVMExpressionNode indexNode;
+                        if (indexType == PrimitiveType.I32) {
+                            indexNode = nodeFactory.createLiteral(1, PrimitiveType.I32);
+                        } else if (indexType == PrimitiveType.I64) {
+                            indexNode = nodeFactory.createLiteral(1L, PrimitiveType.I64);
+                        } else {
+                            throw new AssertionError(indexType);
+                        }
+                        currentAddress = createElementPointer(nodeFactory, addressOffset, currentType, currentAddress, indexNode, indexType, wasVectorized);
+                        wasVectorized = currentAddress instanceof LLVMVectorizedGetElementPtrNodeGen;
+                    }
+                }
+            }
+
+            return currentAddress;
+        } catch (TypeOverflowException e) {
+            return Type.handleOverflowExpression(e);
         }
     }