+++ sulong/projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/debug/type/LLVMSourceFunctionType.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, Oracle and/or its affiliates.
+ * Copyright (c) 2018, 2020, Oracle and/or its affiliates.
  *
  * All rights reserved.
  *
@@ -29,6 +29,7 @@
  */
 package com.oracle.truffle.llvm.runtime.debug.type;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
@@ -97,5 +98,95 @@ public final class LLVMSourceFunctionType extends LLVMSourceType {
     @Override
     public LLVMSourceType getOffset(long newOffset) {
         return this;
+    }
+
+    /**
+     * Helper class used to carry information about function argument locations in source code and
+     * bitcode when they mismatch. Cases where that could happen is when the compiler desugars
+     * structs, examples:
+     *
+     * <pre>
+     * struct Point { double x; double y; };
+     * void func (struct Point p); -> void func(double px, double py);
+     * </pre>
+     */
+    public final class SourceArgumentInformation {
+        private final int bitcodeArgIndex;
+        private final int sourceArgIndex;
+        private final int offset;
+        private final int size;
+
+        /**
+         * @param bitcodeArgIndex Argument location in bitcode.
+         * @param sourceArgIndex Argument location in source code.
+         * @param offset The offset in bits of the bitcode argument in the source code (e.g. in a
+         *            struct).
+         * @param size The size of the argument type in bits.
+         */
+        SourceArgumentInformation(int bitcodeArgIndex, int sourceArgIndex, int offset, int size) {
+            this.bitcodeArgIndex = bitcodeArgIndex;
+            this.sourceArgIndex = sourceArgIndex;
+            this.offset = offset;
+            this.size = size;
+        }
+
+        public long getBitcodeArgIndex() {
+            return this.bitcodeArgIndex;
+        }
+
+        public int getSourceArgIndex() {
+            return this.sourceArgIndex;
+        }
+
+        /**
+         * @return The member offset in bits.
+         */
+        public int getOffset() {
+            return this.offset;
+        }
+
+        /**
+         * @return The argument type size in bits.
+         */
+        public int getSize() {
+            return this.size;
+        }
+    }
+
+    /**
+     * List carrying information for function arguments that mismatch between the source code and
+     * the bitcode. It is expected for arguments that do not have any mismatches to have a null
+     * entry here. This list can be null if there's no debugging information (or no recognizable
+     * debugging information, e.g. due to a missing implementation) associated with the function.
+     */
+    private ArrayList<SourceArgumentInformation> sourceArgumentInformationList;
+
+    /**
+     * Add information for an argument at the location equal to its bitcode location, any arguments
+     * in between shall be set to null (see the
+     * {@link LLVMSourceFunctionType#sourceArgumentInformationList}). Nulls representing
+     * non-mismatching arguments that come after the last mismatching argument are dealt with in
+     * {@link LLVMSourceFunctionType#getSourceArgumentInformation}.
+     */
+    public void attachSourceArgumentInformation(int bitcodeArgIndex, int sourceArgIndex, int offset, int size) {
+        if (sourceArgumentInformationList == null) {
+            sourceArgumentInformationList = new ArrayList<>();
+        }
+
+        int diff = bitcodeArgIndex + 1 - sourceArgumentInformationList.size();
+        assert diff >= 0 || sourceArgumentInformationList.get(bitcodeArgIndex) == null : String.format("sourceArgumentInformation for bitcodeArgIndex %d is already set", bitcodeArgIndex);
+        for (; diff > 0; diff--) {
+            sourceArgumentInformationList.add(null);
+        }
+
+        sourceArgumentInformationList.set(bitcodeArgIndex, new SourceArgumentInformation(bitcodeArgIndex, sourceArgIndex, offset, size));
+    }
+
+    public SourceArgumentInformation getSourceArgumentInformation(int index) {
+        if (sourceArgumentInformationList == null || index >= sourceArgumentInformationList.size()) {
+            return null;
+        }
+
+        return sourceArgumentInformationList.get(index);
     }
 }