+++ sulong/projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/interop/LLVMGetInteropParamNode.java
@@ -27,18 +27,13 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
  * OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+package com.oracle.truffle.llvm.runtime.interop;
 
-#include <stdio.h>
+import com.oracle.truffle.llvm.runtime.nodes.api.LLVMNode;
 
-/* The noinline attribute causes clang to not realize that calling this function with NULL
- * would result in undefined behavior and replacing the call to it in main() with an
- * "unreachable" instruction.
+/**
+ * Node for mapping bitcode function parameters to interop parameters.
  */
-__attribute__((noinline)) void call_and_print(const char *(*fun)()) {
-  printf("%s\n", fun());
-}
-
-int main() {
-  call_and_print(NULL);
-  return 0;
+public abstract class LLVMGetInteropParamNode extends LLVMNode {
+    public abstract Object execute(Object[] arguments);
 }