+++ sulong/projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/interop/LLVMGetInteropPrimitiveParamNode.java
@@ -27,18 +27,40 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
  * OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+package com.oracle.truffle.llvm.runtime.interop;
 
-#include <stdio.h>
+import com.oracle.truffle.api.dsl.Specialization;
+import com.oracle.truffle.llvm.runtime.CommonNodeFactory;
+import com.oracle.truffle.llvm.runtime.interop.access.LLVMInteropType;
+import com.oracle.truffle.llvm.runtime.interop.convert.ForeignToLLVM;
 
-/* The noinline attribute causes clang to not realize that calling this function with NULL
- * would result in undefined behavior and replacing the call to it in main() with an
- * "unreachable" instruction.
+/**
+ * Node for 1:1 mapping of primitive bitcode parameters to interop parameters.
  */
-__attribute__((noinline)) void call_and_print(const char *(*fun)()) {
-  printf("%s\n", fun());
-}
+public abstract class LLVMGetInteropPrimitiveParamNode extends LLVMGetInteropParamNode {
+    public static LLVMGetInteropPrimitiveParamNode create(int index, ForeignToLLVM.ForeignToLLVMType type) {
+        return LLVMGetInteropPrimitiveParamNodeGen.create(index, type);
+    }
 
-int main() {
-  call_and_print(NULL);
-  return 0;
+    public static LLVMGetInteropPrimitiveParamNode create(int index, LLVMInteropType.Value type) {
+        return LLVMGetInteropPrimitiveParamNodeGen.create(index, type);
+    }
+
+    private final int index;
+    @Child ForeignToLLVM toLLVM;
+
+    LLVMGetInteropPrimitiveParamNode(int index, ForeignToLLVM.ForeignToLLVMType type) {
+        this.index = index;
+        this.toLLVM = CommonNodeFactory.createForeignToLLVM(type);
+    }
+
+    LLVMGetInteropPrimitiveParamNode(int index, LLVMInteropType.Value type) {
+        this.index = index;
+        this.toLLVM = CommonNodeFactory.createForeignToLLVM(type);
+    }
+
+    @Specialization
+    Object getParam(Object[] arguments) {
+        return toLLVM.executeWithTarget(arguments[this.index]);
+    }
 }