+++ sulong/projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/interop/access/LLVMInteropType.java
@@ -152,7 +152,7 @@ public abstract class LLVMInteropType implements TruffleObject {
             return new Value(kind, null, size);
         }
 
-        static Value pointer(Structured baseType, long size) {
+        public static Value pointer(Structured baseType, long size) {
             return new Value(ValueKind.POINTER, baseType, size);
         }
 
@@ -339,7 +339,7 @@ public abstract class LLVMInteropType implements TruffleObject {
             return parameterTypes[i];
         }
 
-        public int getParameterLength() {
+        public int getNumberOfParameters() {
             return parameterTypes.length;
         }
     }