+++ sulong/tests/com.oracle.truffle.llvm.tests/src/com/oracle/truffle/llvm/tests/interop/LLVMInteropTest.java
@@ -996,8 +996,8 @@ public class LLVMInteropTest {
     }
 
     @Test
-    public void test073() {
-        try (Runner runner = new Runner("interop073")) {
+    public void testPolyglotGetArg() {
+        try (Runner runner = new Runner("polyglotGetArg")) {
             Assert.assertEquals(42, runner.run());
         }
     }
@@ -1545,7 +1545,7 @@ public class LLVMInteropTest {
     }
 
     private static final Path TEST_DIR = new File(TestOptions.TEST_SUITE_PATH, "interop").toPath();
-    public static final String FILENAME = "O0_MEM2REG." + NFIContextExtension.getNativeLibrarySuffix();
+    public static final String FILENAME = "O1." + NFIContextExtension.getNativeLibrarySuffix();
 
     protected static Map<String, String> getSulongTestLibContextOptions() {
         Map<String, String> map = new HashMap<>();